organize eslint rule assets under eslint-rules
