+++ eslint-rules/eslint-rules.md
@@ -1,0 +1,11 @@
+# ESLint rules
+
+Custom ESLint rule assets live in this directory:
+
+- implementation files (`*.js`)
+- rule docs (`*.md`)
+- rule tests (`*.test.js`)
+
+## Available rules
+
+- [`epic-web/prefer-dispose-in-tests`](./prefer-dispose-in-tests.md)
